Share several new models in export.wiki between stages
